requirements.txt
streamlit
pillow
numpy
scikit-learn
joblib
opencv-python-headless